auxiliary/requirements.txt
exif==1.3.4
Pillow==9.3.0
plum-py==0.7.2
